added feature and steps
