+++ src/test/E2E/steps/admin/AdminAction.java
@@ -1,0 +1,61 @@
+package E2E.steps.admin;
+
+import io.cucumber.java.en.Given;
+import io.cucumber.java.en.Then;
+import io.cucumber.java.en.When;
+
+public class AdminAction {
+    @Given("the admin is on admin page")
+    public void the_admin_is_on_admin_page() {
+        // Write code here that turns the phrase above into concrete actions
+        throw new io.cucumber.java.PendingException();
+    }
+
+    @When("the admin clicks delete user")
+    public void the_admin_clicks_delete_user() {
+        // Write code here that turns the phrase above into concrete actions
+        throw new io.cucumber.java.PendingException();
+    }
+
+    @Then("the user should not appear on the users list")
+    public void the_user_should_not_appear_on_the_users_list() {
+        // Write code here that turns the phrase above into concrete actions
+        throw new io.cucumber.java.PendingException();
+    }
+
+    @When("the admin clicks on posts")
+    public void the_admin_clicks_on_posts() {
+        // Write code here that turns the phrase above into concrete actions
+        throw new io.cucumber.java.PendingException();
+    }
+
+    @When("the admin clicks delete posts")
+    public void the_admin_clicks_delete_posts() {
+        // Write code here that turns the phrase above into concrete actions
+        throw new io.cucumber.java.PendingException();
+    }
+
+    @Then("the post should not appear on the post list")
+    public void the_post_should_not_appear_on_the_post_list() {
+        // Write code here that turns the phrase above into concrete actions
+        throw new io.cucumber.java.PendingException();
+    }
+
+    @When("the admin clicks see comments button")
+    public void the_admin_clicks_see_comments_button() {
+        // Write code here that turns the phrase above into concrete actions
+        throw new io.cucumber.java.PendingException();
+    }
+
+    @When("the admin clicks delete comment")
+    public void the_admin_clicks_delete_comment() {
+        // Write code here that turns the phrase above into concrete actions
+        throw new io.cucumber.java.PendingException();
+    }
+
+    @Then("the comment should not appear on the comments list")
+    public void the_comment_should_not_appear_on_the_comments_list() {
+        // Write code here that turns the phrase above into concrete actions
+        throw new io.cucumber.java.PendingException();
+    }
+}
