+++ src/test/E2E/steps/admin/AdminLogin.java
@@ -1,0 +1,50 @@
+package E2E.steps.admin;
+
+import io.cucumber.java.en.Given;
+import io.cucumber.java.en.Then;
+import io.cucumber.java.en.When;
+
+public class AdminLogin {
+
+    @Given("the user is on the login page")
+    public void the_user_is_on_the_login_page() {
+        // Write code here that turns the phrase above into concrete actions
+        throw new io.cucumber.java.PendingException();
+    }
+
+    @When("the admin enters the username <username>")
+    public void the_admin_enters_the_username_username() {
+        // Write code here that turns the phrase above into concrete actions
+        throw new io.cucumber.java.PendingException();
+    }
+
+    @When("the admin enters the password <password>")
+    public void the_admin_enters_the_password_password() {
+        // Write code here that turns the phrase above into concrete actions
+        throw new io.cucumber.java.PendingException();
+    }
+
+    @Then("the admin should be redirected to the admin page")
+    public void the_admin_should_be_redirected_to_the_admin_page() {
+        // Write code here that turns the phrase above into concrete actions
+        throw new io.cucumber.java.PendingException();
+    }
+
+    @Given("the admin is on the admin page")
+    public void the_admin_is_on_the_admin_page() {
+        // Write code here that turns the phrase above into concrete actions
+        throw new io.cucumber.java.PendingException();
+    }
+
+    @When("the admin clicks log out button")
+    public void the_admin_clicks_log_out_button() {
+        // Write code here that turns the phrase above into concrete actions
+        throw new io.cucumber.java.PendingException();
+    }
+
+    @Then("the admin should be redirected to the log in page")
+    public void the_admin_should_be_redirected_to_the_log_in_page() {
+        // Write code here that turns the phrase above into concrete actions
+        throw new io.cucumber.java.PendingException();
+    }
+}
