+++ src/test/E2E/steps/user/UserAction.java
@@ -1,0 +1,85 @@
+package E2E.steps.user;
+
+import io.cucumber.java.en.Then;
+import io.cucumber.java.en.When;
+
+public class UserAction {
+
+    @When("the user enters something")
+    public void the_user_enters_something() {
+        // Write code here that turns the phrase above into concrete actions
+        throw new io.cucumber.java.PendingException();
+    }
+
+    @When("the user clicks post button")
+    public void the_user_clicks_post_button() {
+        // Write code here that turns the phrase above into concrete actions
+        throw new io.cucumber.java.PendingException();
+    }
+
+    @Then("the post will be displayed at the top of the feed")
+    public void the_post_will_be_displayed_at_the_top_of_the_feed() {
+        // Write code here that turns the phrase above into concrete actions
+        throw new io.cucumber.java.PendingException();
+    }
+
+    @When("the user clicks delete post button")
+    public void the_user_clicks_delete_post_button() {
+        // Write code here that turns the phrase above into concrete actions
+        throw new io.cucumber.java.PendingException();
+    }
+
+    @Then("the post will be deleted")
+    public void the_post_will_be_deleted() {
+        // Write code here that turns the phrase above into concrete actions
+        throw new io.cucumber.java.PendingException();
+    }
+
+    @When("the user clicks like post button")
+    public void the_user_clicks_like_post_button() {
+        // Write code here that turns the phrase above into concrete actions
+        throw new io.cucumber.java.PendingException();
+    }
+
+    @Then("the like count will increment by {int}")
+    public void the_like_count_will_increment_by(Integer int1) {
+        // Write code here that turns the phrase above into concrete actions
+        throw new io.cucumber.java.PendingException();
+    }
+
+    @When("the user clicks unlike comment button")
+    public void the_user_clicks_unlike_comment_button() {
+        // Write code here that turns the phrase above into concrete actions
+        throw new io.cucumber.java.PendingException();
+    }
+
+    @Then("the like count will decrement {int}")
+    public void the_like_count_will_decrement(Integer int1) {
+        // Write code here that turns the phrase above into concrete actions
+        throw new io.cucumber.java.PendingException();
+    }
+
+    @When("the user enters their comment")
+    public void the_user_enters_their_comment() {
+        // Write code here that turns the phrase above into concrete actions
+        throw new io.cucumber.java.PendingException();
+    }
+
+    @When("the user clicks comment button")
+    public void the_user_clicks_comment_button() {
+        // Write code here that turns the phrase above into concrete actions
+        throw new io.cucumber.java.PendingException();
+    }
+
+    @When("the user clicks on delete comment button")
+    public void the_user_clicks_on_delete_comment_button() {
+        // Write code here that turns the phrase above into concrete actions
+        throw new io.cucumber.java.PendingException();
+    }
+
+    @Then("the comment will be deleted")
+    public void the_comment_will_be_deleted() {
+        // Write code here that turns the phrase above into concrete actions
+        throw new io.cucumber.java.PendingException();
+    }
+}
