+++ src/test/E2E/steps/user/UserFollow.java
@@ -1,0 +1,37 @@
+package E2E.steps.user;
+
+import io.cucumber.java.en.Given;
+import io.cucumber.java.en.Then;
+import io.cucumber.java.en.When;
+
+public class UserFollow {
+    @Given("the user is on the other user's profile page")
+    public void the_user_is_on_the_other_user_s_profile_page() {
+        // Write code here that turns the phrase above into concrete actions
+        throw new io.cucumber.java.PendingException();
+    }
+
+    @When("the user click on the follow button")
+    public void the_user_click_on_the_follow_button() {
+        // Write code here that turns the phrase above into concrete actions
+        throw new io.cucumber.java.PendingException();
+    }
+
+    @Then("the user should now be followed and the follow button should change to unfollow")
+    public void the_user_should_now_be_followed_and_the_follow_button_should_change_to_unfollow() {
+        // Write code here that turns the phrase above into concrete actions
+        throw new io.cucumber.java.PendingException();
+    }
+
+    @When("the user clicks on the unfollow button")
+    public void the_user_clicks_on_the_unfollow_button() {
+        // Write code here that turns the phrase above into concrete actions
+        throw new io.cucumber.java.PendingException();
+    }
+
+    @Then("the user should now be unfollowed and the unfollow button should change to follow")
+    public void the_user_should_now_be_unfollowed_and_the_unfollow_button_should_change_to_follow() {
+        // Write code here that turns the phrase above into concrete actions
+        throw new io.cucumber.java.PendingException();
+    }
+}
