+++ src/test/E2E/steps/user/UserLogin.java
@@ -1,0 +1,91 @@
+package E2E.steps.user;
+
+import io.cucumber.java.en.Given;
+import io.cucumber.java.en.Then;
+import io.cucumber.java.en.When;
+
+public class UserLogin {
+    @Given("the user is on login page")
+    public void the_user_is_on_login_page() {
+        // Write code here that turns the phrase above into concrete actions
+        throw new io.cucumber.java.PendingException();
+    }
+
+    @When("the user clicks on register button")
+    public void the_user_clicks_on_register_button() {
+        // Write code here that turns the phrase above into concrete actions
+        throw new io.cucumber.java.PendingException();
+    }
+
+    @When("the user enters username")
+    public void the_user_enters_username() {
+        // Write code here that turns the phrase above into concrete actions
+        throw new io.cucumber.java.PendingException();
+    }
+
+    @When("the user enters password")
+    public void the_user_enters_password() {
+        // Write code here that turns the phrase above into concrete actions
+        throw new io.cucumber.java.PendingException();
+    }
+
+    @When("the user enters first name")
+    public void the_user_enters_first_name() {
+        // Write code here that turns the phrase above into concrete actions
+        throw new io.cucumber.java.PendingException();
+    }
+
+    @When("the user enters last name")
+    public void the_user_enters_last_name() {
+        // Write code here that turns the phrase above into concrete actions
+        throw new io.cucumber.java.PendingException();
+    }
+
+    @When("the user enters email")
+    public void the_user_enters_email() {
+        // Write code here that turns the phrase above into concrete actions
+        throw new io.cucumber.java.PendingException();
+    }
+
+    @When("the user clicks on create register button")
+    public void the_user_clicks_on_create_register_button() {
+        // Write code here that turns the phrase above into concrete actions
+        throw new io.cucumber.java.PendingException();
+    }
+
+    @Then("the user will be redirected to the user profile")
+    public void the_user_will_be_redirected_to_the_user_profile() {
+        // Write code here that turns the phrase above into concrete actions
+        throw new io.cucumber.java.PendingException();
+    }
+
+    @When("the user clicks login button")
+    public void the_user_clicks_login_button() {
+        // Write code here that turns the phrase above into concrete actions
+        throw new io.cucumber.java.PendingException();
+    }
+
+    @Then("the user will be redirected to the homepage")
+    public void the_user_will_be_redirected_to_the_homepage() {
+        // Write code here that turns the phrase above into concrete actions
+        throw new io.cucumber.java.PendingException();
+    }
+
+    @Given("the user is on homepage")
+    public void the_user_is_on_homepage() {
+        // Write code here that turns the phrase above into concrete actions
+        throw new io.cucumber.java.PendingException();
+    }
+
+    @When("the user clicks logout button")
+    public void the_user_clicks_logout_button() {
+        // Write code here that turns the phrase above into concrete actions
+        throw new io.cucumber.java.PendingException();
+    }
+
+    @Then("the user will be redirected to the login page")
+    public void the_user_will_be_redirected_to_the_login_page() {
+        // Write code here that turns the phrase above into concrete actions
+        throw new io.cucumber.java.PendingException();
+    }
+}
