+++ src/test/E2E/steps/user/UserView.java
@@ -1,0 +1,49 @@
+package E2E.steps.user;
+
+import io.cucumber.java.en.Then;
+import io.cucumber.java.en.When;
+
+public class UserView {
+
+    @When("the user enters other user's first name on the search bar")
+    public void the_user_enters_other_user_s_first_name_on_the_search_bar() {
+        // Write code here that turns the phrase above into concrete actions
+        throw new io.cucumber.java.PendingException();
+    }
+
+    @When("the user clicks search button")
+    public void the_user_clicks_search_button() {
+        // Write code here that turns the phrase above into concrete actions
+        throw new io.cucumber.java.PendingException();
+    }
+
+    @Then("the user should see the searched user's name and a link to their profile")
+    public void the_user_should_see_the_searched_user_s_name_and_a_link_to_their_profile() {
+        // Write code here that turns the phrase above into concrete actions
+        throw new io.cucumber.java.PendingException();
+    }
+
+    @When("the user clicks on the My Profile button")
+    public void the_user_clicks_on_the_my_profile_button() {
+        // Write code here that turns the phrase above into concrete actions
+        throw new io.cucumber.java.PendingException();
+    }
+
+    @Then("the user should be redirect to their profile page")
+    public void the_user_should_be_redirect_to_their_profile_page() {
+        // Write code here that turns the phrase above into concrete actions
+        throw new io.cucumber.java.PendingException();
+    }
+
+    @When("the user clicks on the searched person")
+    public void the_user_clicks_on_the_searched_person() {
+        // Write code here that turns the phrase above into concrete actions
+        throw new io.cucumber.java.PendingException();
+    }
+
+    @Then("the user should be redirected to the searched user's profile page")
+    public void the_user_should_be_redirected_to_the_searched_user_s_profile_page() {
+        // Write code here that turns the phrase above into concrete actions
+        throw new io.cucumber.java.PendingException();
+    }
+}
